# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "三方比价支出依据扫描工具",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/files": {
      "post": {
        "tags": [
          "文件导入"
        ],
        "summary": "Upload File",
        "description": "上传供应商文件，异步解析，返回 task_id 和 file_id",
        "operationId": "upload_file_api_projects__project_id__files_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_projects__project_id__files_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "文件导入"
        ],
        "summary": "List Files",
        "description": "获取项目的所有已导入文件",
        "operationId": "list_files_api_projects__project_id__files_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierFileResponse"
                  },
                  "title": "Response List Files Api Projects  Project Id  Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{file_id}/confirm-supplier": {
      "put": {
        "tags": [
          "文件导入"
        ],
        "summary": "Confirm Supplier",
        "description": "确认供应商名称",
        "operationId": "confirm_supplier_api_files__file_id__confirm_supplier_put",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{file_id}": {
      "delete": {
        "tags": [
          "文件导入"
        ],
        "summary": "Delete File",
        "description": "删除未确认供应商的文件及其关联数据",
        "operationId": "delete_file_api_files__file_id__delete",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete File Api Files  File Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/tables": {
      "get": {
        "tags": [
          "文件导入"
        ],
        "summary": "List Tables",
        "description": "获取项目的所有解析表格",
        "operationId": "list_tables_api_projects__project_id__tables_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RawTableResponse"
                  },
                  "title": "Response List Tables Api Projects  Project Id  Tables Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tables/{table_id}/toggle-selection": {
      "put": {
        "tags": [
          "文件导入"
        ],
        "summary": "Toggle Table Selection",
        "description": "切换表格参与比价状态",
        "operationId": "toggle_table_selection_api_tables__table_id__toggle_selection_put",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableToggleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_api_projects__project_id__files_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_projects__project_id__files_post"
      },
...
      "FileUploadResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "supplier_name_guess": {
            "type": "string",
            "title": "Supplier Name Guess"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "task_id",
          "supplier_name_guess"
        ],
        "title": "FileUploadResponse",
        "description": "文件上传响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RawTableResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "supplier_file_id": {
            "type": "string",
            "title": "Supplier File Id"
          },
          "table_index": {
            "type": "integer",
            "title": "Table Index"
          },
          "sheet_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Name"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "column_count": {
            "type": "integer",
            "title": "Column Count"
          },
          "raw_data": {
            "title": "Raw Data"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected"
          },
          "supplier_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Name"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "supplier_confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Confirmed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "supplier_file_id",
          "table_index",
          "row_count",
          "column_count",
          "raw_data",
          "selected"
        ],
        "title": "RawTableResponse",
        "description": "API 返回的原始表格信息"
      },
...
      "SupplierConfirmRequest": {
        "properties": {
          "supplier_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Supplier Name"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "supplier_name",
          "project_id"
        ],
        "title": "SupplierConfirmRequest",
        "description": "确认供应商名称请求"
      },
      "SupplierFileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "supplier_name": {
            "type": "string",
            "title": "Supplier Name"
          },
          "supplier_confirmed": {
            "type": "boolean",
            "title": "Supplier Confirmed"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "recognition_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recognition Mode"
          },
          "imported_at": {
            "type": "string",
            "title": "Imported At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "supplier_name",
          "supplier_confirmed",
          "original_filename",
          "file_path",
          "file_type",
          "imported_at"
        ],
        "title": "SupplierFileResponse",
        "description": "API 返回的文件信息"
      },
...
      "TableToggleRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "TableToggleRequest",
        "description": "表格选择切换请求"
      },
      "TableToggleResponse": {
        "properties": {
          "table_id": {
            "type": "string",
            "title": "Table Id"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected"
          }
        },
        "type": "object",
        "required": [
          "table_id",
          "selected"
        ],
        "title": "TableToggleResponse",
        "description": "表格选择切换响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}